ChartNarrator/requirements.txt
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
transformers==4.44.0
datasets==2.21.0
accelerate==0.33.0
peft==0.11.1
sentencepiece
evaluate
sacrebleu
rouge-score
pillow
matplotlib
gradio
tqdm
